Moved Matching trait higher in the hierarchy
