+++ be2-scala/src/main/scala/ch/epfl/pop/model/network/Matching.scala
@@ -1,4 +1,4 @@
-package ch.epfl.pop.model.network.method.message.data
+package ch.epfl.pop.model.network
 
 sealed trait Matching {
   // See https://stackoverflow.com/questions/3407032/comparing-string-and-enumeration
